simplified Search component folder
