+++ src/components/index.js
@@ -1,4 +1,4 @@
-export { FilterProvider, FilterContainer } from "./Filter/index";
-export { SearchProvider, Search, useSearch } from "./Search/index";
+export { FilterProvider, FilterContainer } from "./Filter";
+export { SearchProvider, Search, useSearch } from "./Search";
 export { default as Results } from "./Results/Results";
 export { default as Footer } from "./Footer/Footer";
